Use serilog for logging (first pass)
So far I've just hooked Serilog up to Topshelf, replaced the inner workings of the Logger class to use it as well as using it directly in some places. I haven't completely dismantled the Logger class because it adds useful information to the logs, such as request ID and hostname. However, I think there will be better ways to inject this data into the log events using Serilog's support for structured data.

See COMPUTE-9